requirements.txt
streamlit
pandas
pdfplumber
python-dotenv
google-generativeai
google-api-core
